DentalClinic/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
attrs==23.1.0
Babel==2.14.0
Django==5.0
django-filter==23.5
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
djangorestframework-filter==0.2.2
drf-spectacular-sidecar==2023.12.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
packaging==23.2
phonenumberslite==8.13.26
Pillow==10.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.13.2
setuptools==69.0.2
six==1.16.0
sqlparse==0.4.4
swagger-spec-validator==3.0.3
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
